tests: extract Wine exec-server fixture
