+++ codex-rs/core/tests/remote_env_windows/BUILD.bazel
@@ -13,14 +13,13 @@ wine_rust_test(
         "codex-app-server": "//codex-rs/app-server:codex-app-server",
     },
     deps = [
-        "//bazel/rules/testing/wine:wine_test_support",
         "//codex-rs/app-server-protocol",
         "//codex-rs/app-server/tests/common",
         "//codex-rs/core/tests/common",
         "//codex-rs/exec-server",
+        "//codex-rs/exec-server/testing:wine-exec-server-test-support",
         "//codex-rs/features",
         "//codex-rs/protocol",
-        "//codex-rs/utils/cargo-bin",
         "//codex-rs/utils/path-uri",
         "@crates//:anyhow",
         "@crates//:pretty_assertions",
